# dist/api-docs.json
{
	"swagger": "2.0",
	"info" : {
		"version": "1.0.0",
		"title": "Boozy analytics API",
        "description": "API to CRUD user behavior data and carry out Boozy analytics board functions."
	},
	"basePath": "/api",
	"tags" : [
		{
			"name": "General",
			"description": "General API."
		},
		{
			"name": "Users",
			"description": "API handling business owner data."
		},
		{
			"name": "Areas",
			"description": "API handling area data."
		},
		{
			"name": "Businesses",
			"description": "API handling business data."
		},
		{
			"name": "Services",
			"description": "API handling service data."
		},
		{
			"name": "Actions",
			"description": "API handling user action data."
		}
	],
	"schemes":[
		"http"
	],
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"paths":{
		"/":{
			"get":{
				"tags": ["General"],
				"description": "Check out API server status.",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Server is running."
					}
				}
			}
		},
		"/setup":{
			"post":{
				"tags": ["General"],
				"description": "Setup administrator account.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "Body",
						"in": "body",
						"description": "Promised passcode to create administrator role.",
						"required": true,
						"schema":{
							"$ref": "#/definitions/passcode"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Success"
					},
					"401": {
						"description": "Wrong passcode"
					},
					"409": {
						"description": "Admin account is already set up."
					}
				}
			}
		},
		"/secret":{
			"post":{
				"tags": ["General"],
				"description": "Get admin information (for test purpose).",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "Body",
						"in": "body",
						"description": "Promised passcode to access administrator info.",
						"required": true,
						"schema":{
							"$ref": "#/definitions/passcode"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Success Response",
						"schema": {
							"$ref": "#/definitions/admin"
						}
					},
					"401": {
						"description": "Wrong passcode"
					}
				}
			}
		},
		"/authenticate":{
			"post":{
				"tags": ["General"],
				"description": "Authenticate the user.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "Body",
						"in": "body",
						"description": "Log in credentials.",
						"required": true,
						"schema": {
							"$ref": "#/definitions/credential"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Authentication",
						"schema": {
							"$ref": "#/definitions/token"
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Invalid credentials"
					}
				}
			}
		},
		"/user":{
			"get":{
				"tags": ["Users"],
				"description": "Get the list of all business owners.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{			
						"name": "x-access-token",
						"in": "header",
						"description": "Access token with admin privilege.",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"201": {
						"description": "Success Response",
						"schema": {
							"$ref": "#/definitions/userlist"
						}
					},
					"401": {
						"description": "Not Authorized."
					}
				}
			},
			"post":{
				"tags": ["Users"],
				"description": "Add a new user.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "details",
						"in": "body",
						"description": "User information",
						"required": true,
						"schema": {
							"$ref": "#/definitions/user"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Creation"
					},
					"403": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Not Authorized"
					}
				}
			}
		},
		"/user/{id}":{
			"put":{
				"tags": ["Users"],
				"description": "Change user information.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{			
						"name": "x-access-token",
						"in": "header",
						"description": "Access token with admin privilege.",
						"required": true,
						"type": "string"
					},
					{
						"name": "id",
						"in": "path",
						"description": "Id of the user to be changed.",
						"required": true,
						"type": "integer"
					},
					{
						"name": "details",
						"in": "body",
						"description": "User information",
						"required": true,
						"schema": {
							"$ref": "#/definitions/user"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Update"
					},
					"403": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Not Authorized"
					}
				}
			}
		},
		"/business": {
			"post": {
				"tags": ["Businesses"],
				"description": "Add a new business.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{			
						"name": "x-access-token",
						"in": "header",
						"description": "Access token with admin privilege.",
						"required": true,
						"type": "string"
					},
					{
						"name": "details",
						"in": "body",
						"description": "Business information",
						"required": true,
						"schema": {
							"$ref": "#/definitions/business"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Creation"
					},
					"403": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Not Authorized"
					}
				}
			},
			"get": {
				"tags": ["Businesses"],
				"description": "Get the list of all businesses.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{			
						"name": "x-access-token",
						"in": "header",
						"description": "Access token with admin privilege.",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"201": {
						"description": "Success Response",
						"schema": {
							"$ref": "#/definitions/businesslist"
						}
					},
					"401": {
						"description": "Not Authorized."
					}
				}
			}
		},
		"/business/{id}":{
			"put":{
				"tags": ["Businesses"],
				"description": "Change business information.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{			
						"name": "x-access-token",
						"in": "header",
						"description": "Access token with admin privilege.",
						"required": true,
						"type": "string"
					},
					{
						"name": "id",
						"in": "path",
						"description": "Id of the business to be changed.",
						"required": true,
						"type": "integer"
					},
					{
						"name": "details",
						"in": "body",
						"description": "Business information",
						"required": true,
						"schema": {
							"$ref": "#/definitions/business"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Update"
					},
					"403": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Not Authorized"
					}
				}
			}
		},
		"/action/{business_code}":{
			"get":{
				"tags": ["Actions"],
				"description": "Get the list of action tracks towards a business.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "business_code",
						"in": "path",
						"description": "Code of the business for which the tracks will be fetched.",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"201": {
						"description": "Success Response",
						"schema": {
							"$ref": "#/definitions/actionlist"
						}
					},
					"403": {
						"description": "Bad Request."
					}
				}
			}
		},
		"/action/{business_code}/{service_type}":{
			"post":{
				"tags": ["Actions"],
				"description": "Record a new action.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "business_code",
						"in": "path",
						"description": "Code of the business to which user action was taken.",
						"required": true,
						"type": "string"
					},
					{
						"name": "service_type",
						"in": "path",
						"description": "Service of the business to which user action was taken (Select from 'brunch' and 'happy').",
						"required": true,
						"type": "string"
					},
					{
						"name": "details",
						"in": "body",
						"description": "Action Type",
						"required": true,
						"schema": {
							"$ref": "#/definitions/actionpost"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Creation"
					},
					"403": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Not Authorized"
					}
				}
			}
		},
		"/area":{
			"post":{
				"tags": ["Areas"],
				"description": "Add a new area.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{			
						"name": "x-access-token",
						"in": "header",
						"description": "Access token with admin privilege.",
						"required": true,
						"type": "string"
					},
					{
						"name": "details",
						"in": "body",
						"description": "Area information",
						"required": true,
						"schema": {
							"$ref": "#/definitions/area"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Creation"
					},
					"403": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Not Authorized"
					}
				}
			},
			"get":{
				"tags": ["Areas"],
				"description": "Get the list of all areas.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{			
						"name": "x-access-token",
						"in": "header",
						"description": "Access token with admin privilege.",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"201": {
						"description": "Success Response",
						"schema": {
							"$ref": "#/definitions/arealist"
						}
					},
					"401": {
						"description": "Not Authorized."
					}
				}
			}
		}
	},
	"definitions": {
		"passcode": {
			"properties": {
				"passcode": {
					"required": true,
					"type": "string",
					"example": "bz_0098_3150"
				}
			}
		},
		"credential": {
			"properties": {
				"email": {
					"required": true,
					"type": "string",
					"example": "admin@test.com"
				},
				"password": {
					"required": true,
					"type": "password",
					"example": "admin"
				}
			}
		},
		"token":{
			"type": "object",
			"required": [
				"success",
				"id",
				"token"
			],
			"properties": {
				"success": {
					"type": "string"
				},
				"id": {
					"type": "integer",
					"description": "User ID"
				},
				"token": {
					"type": "string",
					"description": "Token to be used on privileged requests."
				}
			}
		},
		"admin": {
			"type": "object",
			"properties": {
				"username": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			}
		},
		"user": {
			"type": "object",
			"properties": {
				"username": {
					"type": "string",
					"example": "testuser"
				},
				"email": {
					"type": "string",
					"example": "test@abc.com"
				},
				"password": {
					"type": "password",
					"example": "password"
				},
				"role": {
					"type": "string",
					"example": "owner",
					"description": "Select from 'owner' and 'admin'."
				},
				"business_code": {
					"type": "string",
					"example": "CKQA"
				}

			}
		},
		"userlist": {
			"type" : "object",
			"required" : [
				"success",
				"users"
			],
			"properties": {
				"success": {
					"type": "string"
				},
				"users": {
					"type": "array",
					"items": {
						"user": {
							"type": "object"
						}
					}
				}
			}
		},
		"business": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string",
					"example": "Chicken Republic"
				},
				"area_id": {
					"type": "integer"
				},				
				"code": {
					"type": "string",
					"example": "CKRP"
				}
			}
		},
		"businesslist": {
			"type" : "object",
			"required" : [
				"success",
				"businesses"
			],
			"properties": {
				"success": {
					"type": "string"
				},
				"businesses": {
					"type": "array",
					"items": {
						"business": {
							"type": "object"
						}
					}
				}
			}
		},
		"actionpost":{
			"type": "object",
			"properties": {
				"type": {
					"type": "string",
					"required": true,
					"example": "view",
					"description": "Select from 'view' and 'click'."
				}
			}
		},
		"actionlist":{
			"type" : "object",
			"required" : [
				"success",
				"actions"
			],
			"properties": {
				"success": {
					"type": "string"
				},
				"actions": {
					"type": "array",
					"items": {
						"action": {
							"type": "object"
						}
					}
				}
			}
		},
		"area":{
			"type": "object",
			"properties": {
				"name": {
					"type": "string",
					"example": "New York"
				}
			}
		},
		"arealist":{
			"type" : "object",
			"required" : [
				"success",
				"areas"
			],
			"properties": {
				"success": {
					"type": "string"
				},
				"areas": {
					"type": "array",
					"items": {
						"area": {
							"type": "object"
						}
					}
				}
			}
		}
	}
}